pyproject.toml
[project]
name = "multi_tool_agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.116.1 ",
  "uvicorn>=0.35.0",
  "pydantic>=2.11.7",
  "pydantic-core>=2.33.2",
  "pandas>=2.3.1",
  "openpyxl>=2.3.1",
  "pyyaml>=6.0.2",
  "psycopg2>=2.9.10",
  "sqlalchemy>=2.0",
  "asyncpg>=0.27",
  "ruff==0.12.7",
  "pydantic-settings>=2.10.1",
  "google-adk>=1.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"
fix = true

ignore = ["B008"]

[tool.ruff.format]
preview = true

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []
